Adds NonPlayable abstract class
Refactors Creature to remove the parts that
are only applicable to non-playable creatures.

This was done ahead of adding playable creatures
to the creature hydrator.
