+++ Assets/Creatures/api/NonPlayable.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 98eaea93d0cb4a4580d534a27c10b95a
+timeCreated: 1725505416
