+++ Assets/Creatures/impl/Cerebrian/Cerebrian.cs
@@ -3,7 +3,7 @@ using Creatures.Api;
 
 namespace Creatures.impl
 {
-    public abstract class Cerebrian : Creature, Genera
+    public abstract class Cerebrian : NonPlayable, Genera
     {
         public GeneraType GeneraType { get; } = GeneraType.Cerebrian;
     }
