+++ Assets/Creatures/impl/Frigidariant/Frigidariant.cs
@@ -3,7 +3,7 @@ using Creatures.Api;
 
 namespace Creatures.impl
 {
-    public abstract class Frigidariant : Creature, Genera
+    public abstract class Frigidariant : NonPlayable, Genera
     {
         public GeneraType GeneraType { get; } = GeneraType.Frigidariant;
     }
