+++ Assets/Creatures/impl/Magnarok/Magnarok.cs
@@ -3,7 +3,7 @@ using Creatures.Api;
 
 namespace Creatures.impl
 {
-    public abstract class Magnarok : Creature, Genera
+    public abstract class Magnarok : NonPlayable, Genera
     {
         public GeneraType GeneraType { get; } = GeneraType.Magnarok;
     }
