+++ Assets/Creatures/impl/Venoxarid/Venoxarid.cs
@@ -3,7 +3,7 @@ using Creatures.Api;
 
 namespace Creatures.impl
 {
-    public abstract class Venoxarid : Creature, Genera
+    public abstract class Venoxarid : NonPlayable, Genera
     {
         public GeneraType GeneraType { get; } = GeneraType.Venoxarid;
     }
